requirements.txt
requests~=2.28.2
waitress~=2.1.2
Paste~=3.5.2
Flask~=2.2.2
SQLAlchemy==1.4.44
Werkzeug~=2.2.2
Pillow==9.4.0
numpy~=1.24.2
matplotlib~=3.7.1
shapely~=2.0.1
